+++ changelog.md
@@ -25,6 +25,7 @@
   - Preserves root-level access to `favicon.ico`, `keybase.txt`, `resume-anthony-maki.pdf`, `robots.txt`, and `sitemap.xml`, and sets `Access-Control-Allow-Origin: *` for only them to allow external linking
   - Sets all other static assets to again be behind `/static/`
 - Removes unnecessary code now that serverless is live
+- Serves Material Icons from `static/` instead of from Google
 
 ### Known Issues